requirements.txt
matplotlib>=3.9.0
numpy>=1.26.4
pandas>=2.2.2
pygame==2.6.1
PyOpenGL==3.1.9
torch==2.2.0
